IMPALA-3902: Scheduler improvements for running multiple fragment instances on a single backend
This is an extension of the scheduler and coordinator for multi-threaded execution. It
mainly removes the assumption of having a single fragment instance per host. The approach
taken here is to create parallel data structures and control flow functions, where
necessary, and otherwise to leave the existing single-instance logic in place. The parallel
structures' and functions' names are prefixed with "Mt" to facilitate the enventual clean-up.

Changes to data structures:
- QuerySchedule: per-instance and per-fragment structs with complete execution
  parameters (instead of partially relying on TQueryExecRequest); the per-instance
  execution parameter struct is a child of the per-fragment parameter struct
- explicit fragent id, with range 0..#fragments-1 (instead of relying on an index into
  an array in TQueryExecRequest)

Excluded:
- runtime filter handling
- anything related to RM

Change-Id: I240445199e22f009f4e72fdb8754eb8d77e3d680